Changed View Controller Drastically
It now uses Reality Kit
