+++ AnneFrankAR/AnneFrankAR/ViewController.swift
@@ -14,15 +14,18 @@ import CoreBluetooth
 
 
 
-class ViewController: UIViewController, ARSCNViewDelegate, CBPeripheralDelegate, ARSessionDelegate {
+class ViewController: UIViewController, CBPeripheralDelegate, ARSessionDelegate {
     
     
-    //@IBOutlet var arView: ARView!
+    @IBOutlet var arView: ARView!
     //Connection to the AR Scene View
-    @IBOutlet weak var sceneView: ARSCNView!
+    //@IBOutlet weak var sceneView: ARSCNView!
+    
+    
     
     @IBOutlet weak var menuButton: UIButton!
     
+    /* Old Variables
     var tvPlayer:AVPlayer!
     var made = false;
     private var cbCentralManager: CBCentralManager!
@@ -35,20 +38,32 @@ class ViewController: UIViewController, ARSCNViewDelegate, CBPeripheralDelegate,
     
     var camPos:SCNVector3?
     var camRot:SCNVector3?
-    @IBAction func unwindARView(_ sender:UIStoryboardSegue){}
+     */
+    @IBAction func unwindARView(_ sender:UIStoryboardSegue){
+        let menu = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "MainMenu") as! MainMenu
+        self.navigationController?.pushViewController(menu, animated: false)
+    }
     
     
     override func viewDidLoad() {
         super.viewDidLoad()
-       cbCentralManager = CBCentralManager(delegate: self, queue: nil)
+        //Central manager for beacons is not in use
+       //cbCentralManager = CBCentralManager(delegate: self, queue: nil)
         
-
+        //Old Scene View stuff
+        /*
+         
+         sceneView.delegate = self
+         //show statistics shows framerate in corner, could probably be removed in future
+         sceneView.showsStatistics = true
+         let scene = SCNScene()
+         sceneView.scene = scene
+         
+         sceneView.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(ViewController.handleTap(_:))))
+         //sceneView.addGestureRecognizer(UIPanGestureRecognizer()
+         */
         //init things for scene
-        sceneView.delegate = self
-        //show statistics shows framerate in corner, could probably be removed in future
-        sceneView.showsStatistics = true
-        let scene = SCNScene()
-        sceneView.scene = scene
+        
         
         let menuImage = UIImage(named: "BookBlack")! as UIImage
         menuButton.setImage(menuImage, for: .normal)
@@ -73,216 +88,89 @@ class ViewController: UIViewController, ARSCNViewDelegate, CBPeripheralDelegate,
         self.view.addSubview(button_2)
         self.view.addSubview(button_3)
         
-        sceneView.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(ViewController.handleTap(_:))))
-        //sceneView.addGestureRecognizer(UIPanGestureRecognizer()
+        let config = ARWorldTrackingConfiguration()
+        config.planeDetection = [.horizontal, .vertical]
+        config.environmentTexturing = .automatic
+        
+        if
+            ARWorldTrackingConfiguration.supportsSceneReconstruction(.mesh)
+        {
+            config.sceneReconstruction = .mesh
+        }
+        arView.session.run(config)
+        
+        arView.environment.sceneUnderstanding.options = .occlusion
+        
+
+        let scene = try! Experience.loadBox()
+        
+        // Add the box anchor to the scene
+        arView.scene.anchors.append(scene)
+         
         
     }
     
 
-    
+    //Need to convert these to Reality Kit try! load expereinces
     @objc func buttonAction(sender: UIButton!) {
         
-      setupPortal(portalNum: 1)
+      //setupPortal(portalNum: 1)
     }
     
     @objc func buttonAction2(sender: UIButton!) {
-      setupAnnex()
+      //setupAnnex()
     }
     @objc func buttonAction3(sender: UIButton!)
     {
-        setupChamber()
+        //setupChamber()
     }
   
     
-    @objc func handleTap(_ gesture: UIPanGestureRecognizer){
-        let location = gesture.location(in: self.sceneView)
-        guard let nodeHitTest = self.sceneView.hitTest(location, options: nil).first else{
-            print("no node"); return
-        }
-        let nodeHit = nodeHitTest.node
-        nodeName = nodeHit.name
-        if(nodeName != nil){
-            print(nodeName!)
-            
-            if(nodeName! == "Book")
-            {
-                let bookArtifactScene = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "BurntBookArtifact") as! BurntBookArtifact
-                self.navigationController?.pushViewController(bookArtifactScene, animated: false)
-            }
-            else if(nodeName! == "blendShapesPoster")
-            {
-                let blendShapesPosterScene = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "BlendshapesPoster") as! BlendshapesPoster
-                self.navigationController?.pushViewController(blendShapesPosterScene, animated: false)
-            }
-            if(nodeName! == "Paper")
-            {
-                let quizScene = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "QuizController") as! QuizController
-                self.navigationController?.pushViewController(quizScene, animated: false)
-            }
-            if(nodeName! == "Paper")
-            {
-                let quizScene = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "QuizController") as! QuizController
-                self.navigationController?.pushViewController(quizScene, animated: false)
-            }
-            if(nodeName! == "Cube")
-            {
-                let lessonScene = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "PropagandaLesson") as! PropagandaLesson
-                self.navigationController?.pushViewController(lessonScene, animated: false)
-            }
-        }else{
-            print("nil")
-        }
-    }
-    
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        sceneView.session.delegate = self             // ARSESSION DELEGATE
-        let configuration = ARWorldTrackingConfiguration()
-        sceneView.session.run(configuration)
+        //Reality Kit Setup
+        
     }
     
     override func viewWillDisappear(_ animated: Bool) {
         super.viewWillDisappear(animated)
-        sceneView.session.pause()
+        //arView.session.pause()
     }
     
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
     }
     
-    func setupPortal(portalNum: Int) {
-        //**Rendering order must be above 100 in order to be properly hidden in the portal
-        
-        //node is the spawn point of everything in the scene
-        let node = SCNNode()
-        
-        
-        node.position = camPos!
-        node.eulerAngles = camRot!
-        
-        // broke some of the code out to another function
-        buildRoom(num: portalNum, node: node)
-        
-        if(portalNum == 1) {
-            
-            let fileURL = URL(fileURLWithPath: Bundle.main.path(forResource:"videoplayback", ofType: "mp4")!)
-            tvPlayer = AVPlayer(url: fileURL)
-            
-            // add video to plane
-            let tvGeo = SCNPlane(width:1.6,height: 0.9)
-            tvGeo.firstMaterial?.diffuse.contents = tvPlayer
-            tvGeo.firstMaterial?.isDoubleSided = true
-            
-            // create node containing plane
-            let tvNode = SCNNode(geometry: tvGeo)
-            tvNode.position.z = -0.9
-            
-            // play video
-            node.addChildNode(tvNode)
-            tvPlayer.play()
-            
-            // add subscene stuff
-            let subScene = SCNScene(named: "art.scnassets/Scenes/PropagandaScene.scn")!
-            
-            // more subscene stuff
-            let quiz = subScene.rootNode.childNode(withName: "TextualRhetoric", recursively: true)!
-            quiz.position = SCNVector3.init(0.50, -0.5, -0.75)
-            quiz.eulerAngles = SCNVector3.init(0, 0, 0)
-            
-            // subscene subscene subscene
-            let propagandaLesson = subScene.rootNode.childNode(withName: "cubeydude", recursively: true)!
-            propagandaLesson.position = SCNVector3.init(-0.50, -0.5, -0.75)
-            
-            let germ = subScene.rootNode.childNode(withName: "Cylinder2", recursively: true)!
-            germ.position = SCNVector3.init(-0.75, -0.1, (-length / 2) + width * 3)
-            germ.eulerAngles = SCNVector3.init(0, -90.0.degreesToRadians, 0)
-            
-            let germ_2 = subScene.rootNode.childNode(withName: "Cylinder2_2", recursively: true)!
-            germ_2.position = SCNVector3.init(0.75, -0.1, (-length / 2) + width * 3)
-            germ_2.eulerAngles = SCNVector3.init(0, -90.0.degreesToRadians, 0)
-            
-            // <b>s t a n</b>
-            let person = subScene.rootNode.childNode(withName: "stan", recursively: true)!
-            person.position = SCNVector3.init(0, -0.95, -0.75)
-            person.eulerAngles = SCNVector3.init(0, Double.pi / 16, 0)
-            
-            // addng the burnt book
-            let burntBook = subScene.rootNode.childNode(withName: "BurntBook reference", recursively: true)!
-            burntBook.position = SCNVector3.init(0.30, -0.5, -0.75)
-            burntBook.eulerAngles = SCNVector3.init(0, 0, 0)
-            
-            let bsPoster = subScene.rootNode.childNode(withName: "blendShapesPoster", recursively: true)!
-            bsPoster.position = SCNVector3.init(0.90, 0, (length / 10) + width * 2)
-            bsPoster.eulerAngles = SCNVector3.init(0, Double.pi / 2, 0)
-            
-            // add nodes to scene
-            node.addChildNode(propagandaLesson)
-            node.addChildNode(quiz)
-            node.addChildNode(person)
-            node.addChildNode(germ)
-            node.addChildNode(germ_2)
-            node.addChildNode(burntBook)
-            node.addChildNode(bsPoster)
-        }
-        
-        //This is the final step, officially adding node to the scene itself
-        self.sceneView.scene.rootNode.addChildNode(node)
-    }
-    
-    func setupAnnex()
-    {
-        let node = SCNNode()
-        
-        node.position = camPos!
-        node.eulerAngles = camRot!
-        
-        let scene = SCNScene(named: "art.scnassets/Scenes/AnneFrankAnnex.scn")!
-        
-        node.addChildNode(scene.rootNode)
-        
-        self.sceneView.scene.rootNode.addChildNode(node)
-        
-    }
-    func setupChamber()
-    {
-        let node = SCNNode()
-        
-        node.position = camPos!
-        node.eulerAngles = SCNVector3(x:Float(Double.pi) / 2, y:0, z:0)
-        node.scale = SCNVector3(x:0.1, y:0.1, z:0.1)
-        
-        let scene = SCNScene(named: "art.scnassets/Scenes/Chamber.scn")!
-        
-        node.addChildNode(scene.rootNode)
-        
-        self.sceneView.scene.rootNode.addChildNode(node)
-    }
+   
     
     func session(_ session: ARSession, didUpdate frame: ARFrame) {
-        let transform = frame.camera.transform
-        let rotation = frame.camera.eulerAngles
-        
-        let position = transform.columns.3
+        //Old Scene Kit Setup
+        //let transform = frame.camera.transform
+        //let rotation = frame.camera.eulerAngles
+        //let position = transform.columns.3
         //print(position.x, position.y, position.z)
-        // UPDATING
-        camPos = SCNVector3(position.x, position.y, position.z)
-        camRot = SCNVector3(rotation.x, rotation.y, rotation.z+1.5)
+        //camPos = SCNVector3(position.x, position.y, position.z)
+        //camRot = SCNVector3(rotation.x, rotation.y, rotation.z+1.5)
     
     }
     
     func sessionWasInterrupted(_ session: ARSession) {
         // Inform the user that the session has been interrupted, for example, by presenting an overlay
-        
+        print("interrupted")
     }
     
     func sessionInterruptionEnded(_ session: ARSession) {
         // Reset tracking and/or remove existing anchors if consistent tracking is required
-        
+        print("Interruption Ended, anchors may reset")
     }
     
     
 }
 
+
+
+//Not using beacons anymore
+/*
 extension ViewController :  CBCentralManagerDelegate {
 
     
@@ -359,3 +247,161 @@ func centralManagerDidUpdateState(_ central: CBCentralManager) {
   
     }
 }
+*/
+
+//Not using Gestures anymore, using reality kit
+/*
+@objc func handleTap(_ gesture: UIPanGestureRecognizer){
+    let location = gesture.location(in: self.sceneView)
+    guard let nodeHitTest = self.sceneView.hitTest(location, options: nil).first else{
+        print("no node"); return
+    }
+    let nodeHit = nodeHitTest.node
+    nodeName = nodeHit.name
+    if(nodeName != nil){
+        print(nodeName!)
+        
+        if(nodeName! == "Book")
+        {
+            let bookArtifactScene = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "BurntBookArtifact") as! BurntBookArtifact
+            self.navigationController?.pushViewController(bookArtifactScene, animated: false)
+        }
+        else if(nodeName! == "blendShapesPoster")
+        {
+            let blendShapesPosterScene = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "BlendshapesPoster") as! BlendshapesPoster
+            self.navigationController?.pushViewController(blendShapesPosterScene, animated: false)
+        }
+        if(nodeName! == "Paper")
+        {
+            let quizScene = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "QuizController") as! QuizController
+            self.navigationController?.pushViewController(quizScene, animated: false)
+        }
+        if(nodeName! == "Paper")
+        {
+            let quizScene = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "QuizController") as! QuizController
+            self.navigationController?.pushViewController(quizScene, animated: false)
+        }
+        if(nodeName! == "Cube")
+        {
+            let lessonScene = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "PropagandaLesson") as! PropagandaLesson
+            self.navigationController?.pushViewController(lessonScene, animated: false)
+        }
+    }else{
+        print("nil")
+    }
+}
+ */
+
+//Old Scene Kit way of doing things
+/*
+func setupPortal(portalNum: Int) {
+    //**Rendering order must be above 100 in order to be properly hidden in the portal
+    
+    //node is the spawn point of everything in the scene
+    let node = SCNNode()
+    
+    
+    node.position = camPos!
+    node.eulerAngles = camRot!
+    
+    // broke some of the code out to another function
+    buildRoom(num: portalNum, node: node)
+    
+    if(portalNum == 1) {
+        
+        let fileURL = URL(fileURLWithPath: Bundle.main.path(forResource:"videoplayback", ofType: "mp4")!)
+        tvPlayer = AVPlayer(url: fileURL)
+        
+        // add video to plane
+        let tvGeo = SCNPlane(width:1.6,height: 0.9)
+        tvGeo.firstMaterial?.diffuse.contents = tvPlayer
+        tvGeo.firstMaterial?.isDoubleSided = true
+        
+        // create node containing plane
+        let tvNode = SCNNode(geometry: tvGeo)
+        tvNode.position.z = -0.9
+        
+        // play video
+        node.addChildNode(tvNode)
+        tvPlayer.play()
+        
+        // add subscene stuff
+        let subScene = SCNScene(named: "art.scnassets/Scenes/PropagandaScene.scn")!
+        
+        // more subscene stuff
+        let quiz = subScene.rootNode.childNode(withName: "TextualRhetoric", recursively: true)!
+        quiz.position = SCNVector3.init(0.50, -0.5, -0.75)
+        quiz.eulerAngles = SCNVector3.init(0, 0, 0)
+        
+        // subscene subscene subscene
+        let propagandaLesson = subScene.rootNode.childNode(withName: "cubeydude", recursively: true)!
+        propagandaLesson.position = SCNVector3.init(-0.50, -0.5, -0.75)
+        
+        let germ = subScene.rootNode.childNode(withName: "Cylinder2", recursively: true)!
+        germ.position = SCNVector3.init(-0.75, -0.1, (-length / 2) + width * 3)
+        germ.eulerAngles = SCNVector3.init(0, -90.0.degreesToRadians, 0)
+        
+        let germ_2 = subScene.rootNode.childNode(withName: "Cylinder2_2", recursively: true)!
+        germ_2.position = SCNVector3.init(0.75, -0.1, (-length / 2) + width * 3)
+        germ_2.eulerAngles = SCNVector3.init(0, -90.0.degreesToRadians, 0)
+        
+        // <b>s t a n</b>
+        let person = subScene.rootNode.childNode(withName: "stan", recursively: true)!
+        person.position = SCNVector3.init(0, -0.95, -0.75)
+        person.eulerAngles = SCNVector3.init(0, Double.pi / 16, 0)
+        
+        // addng the burnt book
+        let burntBook = subScene.rootNode.childNode(withName: "BurntBook reference", recursively: true)!
+        burntBook.position = SCNVector3.init(0.30, -0.5, -0.75)
+        burntBook.eulerAngles = SCNVector3.init(0, 0, 0)
+        
+        let bsPoster = subScene.rootNode.childNode(withName: "blendShapesPoster", recursively: true)!
+        bsPoster.position = SCNVector3.init(0.90, 0, (length / 10) + width * 2)
+        bsPoster.eulerAngles = SCNVector3.init(0, Double.pi / 2, 0)
+        
+        // add nodes to scene
+        node.addChildNode(propagandaLesson)
+        node.addChildNode(quiz)
+        node.addChildNode(person)
+        node.addChildNode(germ)
+        node.addChildNode(germ_2)
+        node.addChildNode(burntBook)
+        node.addChildNode(bsPoster)
+    }
+    
+    //This is the final step, officially adding node to the scene itself
+    self.sceneView.scene.rootNode.addChildNode(node)
+}
+
+func setupAnnex()
+{
+    let node = SCNNode()
+    
+    node.position = camPos!
+    node.eulerAngles = camRot!
+    
+    let scene = SCNScene(named: "art.scnassets/Scenes/AnneFrankAnnex.scn")!
+    
+    node.addChildNode(scene.rootNode)
+    
+    self.sceneView.scene.rootNode.addChildNode(node)
+    
+}
+func setupChamber()
+{
+    let node = SCNNode()
+    
+    node.position = camPos!
+    node.eulerAngles = SCNVector3(x:Float(Double.pi) / 2, y:0, z:0)
+    node.scale = SCNVector3(x:0.1, y:0.1, z:0.1)
+    
+    let scene = SCNScene(named: "art.scnassets/Scenes/Chamber.scn")!
+    
+    node.addChildNode(scene.rootNode)
+    
+    self.sceneView.scene.rootNode.addChildNode(node)
+}
+
+      */
+ 
+ */
